+++ cspell.json
@@ -1,6 +1,6 @@
 {
   "version": "0.2",
-  "ignorePaths": ["**/*.ts"],
+  "ignorePaths": ["**/*.ts", "**/*.js", "**/dist"],
   "dictionaryDefinitions": [],
   "dictionaries": [],
   "words": ["strapi", "sami", "ami"],
