+++ frontend/components/content/home/ImpactSection.tsx
