+++ frontend/components/content/home/ProjectSummary.tsx
