+++ frontend/components/content/home/missionStatement.tsx
