+++ frontend/components/content/resources/index.tsx
