+++ frontend/components/content/resources/resourceCard.tsx
