+++ frontend/pages/home.tsx
@@ -6,13 +6,13 @@ import { serverQuery } from "lib/graphql";
 import type { HomeContentQuery, ProjectType } from "../graphql/generated";
 import { HomeContentDocument } from "../graphql/generated";
 import { HeroImageComponent } from "components/common/heroImage";
-import { MissionStatementComponent } from "components/pages/home/missionStatement";
+import { MissionStatementComponent } from "components/content/home/missionStatement";
 import PageSection from "components/layout/pageSection";
-import { ProjectSummaryComponent } from "components/pages/home/ProjectSummary";
-import { GetInvolvedComponent } from "components/pages/home/getInvolved";
-import { ImpactSectionComponent } from "components/pages/home/ImpactSection";
-import { BlogCardComponent } from "components/pages/blog-post/blogCard";
-import { DonateSummary } from "components/pages/home/donateSummary";
+import { ProjectSummaryComponent } from "components/content/home/ProjectSummary";
+import { GetInvolvedComponent } from "components/content/home/getInvolved";
+import { ImpactSectionComponent } from "components/content/home/ImpactSection";
+import { BlogCardComponent } from "components/content/blog-post/blogCard";
+import { DonateSummary } from "components/content/home/donateSummary";
 import Image from "next/image";
 import { useBreakpointQuery } from "lib/breakpoints";
