--- frontend/types/about.ts
@@ -1,5 +1,0 @@
-import type { About } from "../graphql/generated";
-
-export interface IAbout extends About {
-  id: string;
-}
