+++ scripts/src/commands/docker/build.ts
@@ -5,7 +5,6 @@ import execa from "execa";
 import { loadEnv } from "../../utils";
 import { PATHS } from "../../paths";
 import packageJson from "../../../../package.json";
-import { spawnSync } from "child_process";
 
 // version number to tag base image with
 export const BASE_TAG = packageJson.version;
