+++ scripts/src/commands/strapi/typesGenerate.ts
@@ -15,14 +15,17 @@ const program = new Command("types:generate");
 export default program
   .description("Generate types")
   .option("-o --out-dir <string>", "Directory to output type defintions")
-  .action(async (options: Partial<IProgramOptions>) => {
-    const optionsWithDefaults: IProgramOptions = {
-      outDir: PATHS.sharedDir,
-      ...options,
-    };
-    generateQueryTypes();
-    // generateSchemaTypes(optionsWithDefaults);
-  });
+  .action(async () =>
+    // options: Partial<IProgramOptions>
+    {
+      // const optionsWithDefaults: IProgramOptions = {
+      //   outDir: PATHS.sharedDir,
+      //   ...options,
+      // };
+      generateQueryTypes();
+      // generateSchemaTypes(optionsWithDefaults);
+    }
+  );
 
 /***************************************************************************************
  * Main Methods
